+++ xarray/core/duck_array_ops.py
@@ -501,12 +501,14 @@ def last(values, axis, skipna=None):
     return take(values, -1, axis=axis)
 
 
-def rolling_window(array, axis, window, center, fill_value):
+def rolling_window(array, axis, window, center, fill_value, mode):
     """
     Make an ndarray with a rolling window of axis-th dimension.
     The rolling dimension will be placed at the last dimension.
     """
     if isinstance(array, dask_array_type):
-        return dask_array_ops.rolling_window(array, axis, window, center, fill_value)
+        return dask_array_ops.rolling_window(
+            array, axis, window, center, fill_value, mode
+        )
     else:  # np.ndarray
-        return nputils.rolling_window(array, axis, window, center, fill_value)
+        return nputils.rolling_window(array, axis, window, center, fill_value, mode)