+++ doc/whats-new.rst
@@ -25,7 +25,10 @@ Breaking changes
 
 New Features
 ~~~~~~~~~~~~
-
+- Some boundary options ('edge' | 'reflect' | 'symmetric' | 'wrap') are 
+  newly supported in :py:meth:`~xarray.DataArray.rolling.construct`.
+  (:issue:`2007`)
+  By `Keisuke Fujii <https://github.com/fujiisoup>`_.
 
 Bug fixes
 ~~~~~~~~~